fix(fe): use useIntervalFn
