+++ frontend/src/user/home/components/Carousel.story.vue
@@ -1,17 +1,39 @@
 <script setup lang="ts">
+import dummyImg from '@/common/assets/dummy.png'
+import dummyImg2 from '@/common/assets/skkudingLogo.png'
 import Carousel from './Carousel.vue'
-
-const slides = [
-  'https://picsum.photos/id/1032/900/400',
-  'https://picsum.photos/id/1033/900/400',
-  'https://picsum.photos/id/1037/900/400',
-  'https://picsum.photos/id/1035/900/400',
-  'https://picsum.photos/id/1036/900/400'
-]
 </script>
 
 <template>
   <Story>
-    <Carousel :slides="slides" />
+    <Carousel
+      v-if="$router.currentRoute.value.meta.home"
+      :slides="[
+        {
+          topTitle: '성균관대학교',
+          bottomTitle: '제 1회 청정수컵',
+          sub: '첫 번째 우승자가 될 기회!',
+          img: dummyImg,
+          imgAlt: 'dummy',
+          color: 'green'
+        },
+        {
+          topTitle: 'SKKUDING',
+          bottomTitle: '스꾸딩 23-2 신입부원 모집',
+          sub: '프론트엔드 0명, 백엔드 0명',
+          img: dummyImg2,
+          imgAlt: 'dummy',
+          color: 'black'
+        },
+        {
+          topTitle: 'SKKUDING',
+          bottomTitle: '제 1회 해커톤',
+          sub: '무박 2일, 1등 100만원',
+          img: dummyImg2,
+          imgAlt: 'dummy',
+          color: 'white'
+        }
+      ]"
+    />
   </Story>
 </template>
